% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.adria-package.R
\docType{package}
\name{bankr.adria-package}
\alias{bankr.adria}
\alias{bankr.adria-package}
\title{bankr.adria: Reads and Tidy Bank Statements from Banca Adria}
\description{
Reads and Tidy Bank Statements from Banca Adria.
}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea.ranzato01@gmail.com}

}
\keyword{internal}
