% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_cbi}
\alias{tidy_cbi}
\title{Tidy CBI}
\usage{
tidy_cbi(x)
}
\arguments{
\item{x}{A character vector as returned by \code{\link[=read_cbi]{read_cbi()}}.}
}
\value{
A data frame
}
\description{
Tidy CBI
}
